Add a build-dep blacklist hook example
The hook, written in clojure-py, checks whether the package to upload
has backlsited build-dependencies, and throws an error in that
case. If the uploaded changes file does not refer to a dsc, it is
silently ignored.

The list of packages to blacklist as build-deps can be configured with
the 'bd-blacklist' key in the profile, where the value is an array of
package names:

 {
   "bd-blacklist": ["dpatch", "cdbs"]
 }

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>