Fail with grace when given a bad hook path.
This is in an attempt to solve #696659 for Moritz Mühlenhoff. Thanks for
the report, Moritz.

This problem presented it's self when the old dput security block
instructed dput to run the security script. Since we don't provide that
script, we shall fail with grace when the script doesn't exist.

Signed-off-by: Paul Tagliamonte <paultag@debian.org>