Add a new 'required-fields' hook
Some distributions place policy requirements on metadata that is
uploaded. For example, Ubuntu's Stable Release Updates require each
upload to close at least one Launchpad bug, as indicated by the
Launchpad-Bugs-Fixed item in the changes file.

Add a hook which enforces that uploads to the given suites must have the
given fields present and non-empty, and add the above-mentioned item to
the default Ubuntu profile.