Ignore DistroDataOutdated when removing devel() distros
If Ubuntu released a new version and distro-info-data was not updated
with a new development version, distro_info.UbuntuDistroInfo().devel()
will throw DistroDataOutdated. When testing with test_upload.py, this
results in a broken unit test, cause distro_info.py uses the current
date, as shown in #927467. As we want to remove the development version
from the supported list anyhow, we can just ignore the error.